Started work on modularizing
