+++ .gitignore
@@ -29,3 +29,4 @@ Temporary Items
 out
 dev
 __pycache__
+.venv
